minor refactoring and removed unused imports
