+++ jolt-core/src/main/java/com/bazaarvoice/jolt/templatr/spec/TemplatrCompositeSpec.java
@@ -31,8 +31,8 @@ import com.bazaarvoice.jolt.common.tree.ArrayMatchedElement;
 import com.bazaarvoice.jolt.common.tree.MatchedElement;
 import com.bazaarvoice.jolt.common.tree.WalkedPath;
 import com.bazaarvoice.jolt.exception.SpecException;
-import com.bazaarvoice.jolt.common.DataType;
-import com.bazaarvoice.jolt.common.OpMode;
+import com.bazaarvoice.jolt.templatr.DataType;
+import com.bazaarvoice.jolt.templatr.OpMode;
 import com.bazaarvoice.jolt.templatr.TemplatrSpecBuilder;
 
 import java.util.ArrayList;
