+++ jolt-core/src/main/java/com/bazaarvoice/jolt/templatr/spec/TemplatrDefaultLeafSpec.java
@@ -19,7 +19,7 @@ package com.bazaarvoice.jolt.templatr.spec;
 import com.bazaarvoice.jolt.common.tree.MatchedElement;
 import com.bazaarvoice.jolt.common.tree.WalkedPath;
 import com.bazaarvoice.jolt.exception.SpecException;
-import com.bazaarvoice.jolt.common.OpMode;
+import com.bazaarvoice.jolt.templatr.OpMode;
 
 import java.util.List;
 import java.util.Map;
