+++ jolt-core/src/main/java/com/bazaarvoice/jolt/templatr/spec/TemplatrSpec.java
@@ -29,7 +29,7 @@ import com.bazaarvoice.jolt.common.tree.MatchedElement;
 import com.bazaarvoice.jolt.common.tree.WalkedPath;
 import com.bazaarvoice.jolt.exception.SpecException;
 import com.bazaarvoice.jolt.exception.TransformException;
-import com.bazaarvoice.jolt.common.OpMode;
+import com.bazaarvoice.jolt.templatr.OpMode;
 
 import java.util.List;
 import java.util.Map;
